requirements.txt
fastapi
uvicorn
langgraph
langchain-core
langchain-google-genai
tavily
supabase
websockets
python-dotenv
pydantic
